[otnstester] adds initial otnstester tests (#88)
This commit adds the otnstester and some initial tests.

Background:

In Golang, only Go tests executed as go test are able to collect code
coverage. Because of it, code coverage of pyOTNS tests can not be
measured by Codecov.io.

The otnstester wraps the whole OTNS program into a Go test so that we
can build tests and collect code coverage data for all OTNS
functionalities.