[stress-test] relax ping test to allow collision-avoidance delay (#257)
This commit relaxes the check in `ping_latency.py` stress test to 300
msec (from 100 msec). This is to accommodate the "collision avoidance
delay mechanism" which is intended to help prevent self-interference
This mechanism applies a delay after a successful frame tx to a
neighbor which is expected to forward the frame before any next frame
tx.