Feature: add simulated-external-hosts support (#157)
This feature adds support for simulated external IPv6 hosts/servers, that a simulated OT node can communicate with. Traffic to the external simulated host (or 'simhost') is routed to a Thread BR (if it publishes the correct routes on the Thread Network), then sent to the OTNS simulator. OTNS will determine a matching simhost (if any) and will use address/port translation to map the IP datagram to a localhost designated port. On this local port, a particular service/server can be started by the user. For example, a CoAP server or HTTP server.

This feature also serves as preparation for future addition of CCM commissioning of simulated OT nodes. In such case, a Thread BR will "relay" commissioning traffic to an external/remote server like a cBRSKI Registrar.

Includes the following commits:

* [cli] introduction of 'host' CLI command
* [pylibs] Add CCM-commissioning unit test (not actively tested for the moment - preparation for future functions)
* [pylibs] extend Commissioning unit test to include multi-hop commissioning.
* [dispatcher][pylibs] added HostEvents counter
* [dispatcher][event] added event types for IP/UDP forwarding to from host via simulator
* [pylibs] using aiocoap library for testing external CoAP server, in unit tests

Includes commits for test/workflow/bug fixes and necessary improvements:

* [.github] update workflow to separate out the version-unittests, to avoid this always failing on macos build.
* [pylibs] fix unit test (TMF Coap messages may be sent in parallel of the test coap message)
* [pcap] insert a dummy 802.15.4 frame at start of every PCAP file, to create a t=0 simulation time reference. This allows any PCAP entry to be associated to a corresponding radio (trace) log entry with exact (to the microsecond) precision.
* [pcap] fix unit test, related to introduction of t=0 time-reference frame.
* [pylibs] added further comments in partial_dataset case study.
* [ot-rfsim] fix macos build error (due to warn)
* [ot-rfsim] add C lang definitions for macOS build cpp in ot-rfsim project.
* [ot-rfsim] adding 'a' in front of function call parameters; and minor format fixes in C code; adding comments; removing FIXME items.
* [ot-rfsim] fix build for v1.1, v1.2 for using ot....() platform logging functions.
* [pylibs] fixes to test_commissioning - test sometimes failed.
* [pylibs] test_border_router: narrow test topology to avoid spurious test fails.