Feature: make-pretty for all languages / Updates: Github yml build files and make-pretty applied (#169)
This updates the `make-pretty` script to support all used languages in the project. It's based on the same script in the `openthread` repo. Also make-pretty is applied on all sources.

Additionally, Github yml files are updated to better comply with recommended steps order and to bump setup-go to v5. Some steps are renamed to provide more clarity what it does. And the 'build' test is now split in the Go build part and the OT nodes build part. These 2 parts each require a separate matrix (since the nodes build doesn't involve Go, and the Go build doesn't involve C/C++ )

* [ot-rfsim] make-pretty clang-tidy
* [ot-rfsim] make-pretty clang-format
* [all] make-pretty markdown .md
* [ot-rfsim] make-pretty shell
* [pylibs][script] make-pretty python (excluding generates files in proto/)
* [.github] Updated Github yml workflow names; split OTNS and Nodes builds in build.yml; bring checkout before setup-go to avoid warnings. I.e. apply standard order as recommended.
* [script] moved cmake and ninja to requires packages -> needed now with Lint requiring a node-build.