[script] Check for minimum Go/Python3 versions
Adds checks for Go and Python3 minimum versions in the bootstrap script. If a version is too low, the script will stop. For Go, it will check if default golang package installation would retrieve a usable version. If it does, the package is installed. If it doesn't, the script will stop.  Also fixes exclude dirs in common.sh.