[docker] fix playground Docker port binding and enable manual GUI start (#653)
This fixes the TCP port issue in the playground Docker image, by setting
the socket binding host to 0.0.0.0. To address the issue that the web GUI
can't be auto-started, a link is displayed which the user can click to open
the web GUI. In the README, this Docker image is documented as the
quickest way to get started (try-out).