[ot-rfsim][dispatcher] enable TCAT Commissioner access to OTNS node over UDP port (#685)
This adds support for TCAT commissioning of OT nodes using the Python TCAT Commissioner. UDP ports are used
to access, so that TCAT/BLE Commissioner in UDP simulation mode can be directly used. It uses port 10000 plus
gNodeId.

It also fixes a bug that caused Thread radio operation to be unavailable when TCAT advertisements are sent by a node.
To resolve this, a new event type 'schedule-node' is introduced that allows the OT-RFSIM platform code to schedule
a new processing event for itself at any future time. This event type is now also internally used by OTNS to
schedule future 'waking' of the node, which more cleanly separates the node scheduling from the alarm-management
which uses the 'alarm' event with a different semantics.