[cli][ot-rfsim][simulation][types] Add 'bitrate' radio parameter / Enable defaults for radio parameters
This adds a 'bitrate' rfsim parameter for setting PHY bitrate per node. It determines the time duration
of a packet transmission.

The command 'rfsim default' is added to set any default rfsim parameters (such as bitrate or others) for
every new node that is created from then on. This is useful to set at OTNS startup (e.g. in a script),
to ensure that all nodes speak the same bitrate or use the same value for Rx sensitivity or CCA
threshold, for example. A unit test is added that tries these default parameters.